refactor: move states to state mod
