+++ vrc-get-gui/src/utils.rs
@@ -1,4 +1,5 @@
-use crate::commands::SettingMutRef;
+use crate::state::*;
+
 use std::borrow::Cow;
 use std::path::{Path, PathBuf};
 use tauri::api::dir::is_dir;
